requirements.txt
colorlog
requests
pyyaml
python-telegram-bot
